requirements.txt
fastapi
uvicorn
requests
groq
python-dotenv
rich
mcp[cli]>=1.0.0